add offset options to input file. Separate input file handling and CLI code.
